cvar: Proof of concept for "internal variables" as a subset of cvar_t
The idea is to create a dividing line between internal variables and
their end-user configurability. This gives more control to the game dev,
where they can decide which engine variables they want configurable with
what ever name they choose. So far it works with sv_cheats.

This allows the engine to be in a more barebones state, not imposing on
the game dev as far as cvars are concerned. It makes cross-engine
compatibility more feasible by allowing variables to adopt different
names to satisfy MenuQC settings menus or what have you.