physics: fix a subtle prediction regression
In 3d8f1ca7b142c0a9fd8c3b4a494b62dcce39ab71 a horizontal player speed
calculation was changed to include the vertical velocity and to no
longer be the speed, by use of the wrong macro.
This patch reverts those changes and adds a macro for this calculation.

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>