+svPortRange, +clPortRange
this sets more static port range to use for client and server.
you can switch that  PortStart + PortEnd
or change values for it was like more clear for people and actually, if they run it run from 1 host or network with NAT, then client and server should not use same port by any occasion - it may fail then/ you can think about it.