cl_screen: Use Sys_DirtyTime while rendering loading screen
This is a partial revert of 40cdb054026fe5acb26c79adabe3db55bee282c2.

We can't rely on `host.realtime`/`dirtytime`, because those are only
updated on calls to `Sys_Frame` (which is Darkplaces' event loop).
When the client receives certain packets, like the server info packet
or the sign-on packets, the client will synchronously load assets like
sounds, models etc. and will "push" a loading screen. Multiple
pushes/pops can happen without the event loop being pumped, which makes
things behave like the game is "frozen". To still render the loading
screen in this state, we grab the current time in an ad-hoc manner
again. This makes the loading screen progress.

The correct fix here is to avoid doing long work in the main thread
and either asynchronously load models or to incrementally load them,
but that would require significant changes.